requirements.txt
allure-pytest==2.8.11
allure-python-commons==2.8.11
Appium-Python-Client==0.50
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
enum34==1.1.9
get==2019.4.13
idna==2.8
lxml==4.5.0
more-itertools==8.2.0
namedlist==1.7
packaging==19.2
pbr==5.4.4
pluggy==0.13.1
poseidon==1.0.1.dev1+g49d79e1.d20200226
post==2019.4.13
public==2019.4.13
py==1.8.0
pycurl==7.43.0.5
pyDes==2.0.1
pymemcache==3.0.0
pyparsing==2.4.5
pytest==5.3.5
pytest-base-url==1.4.1
pytest-html==3.0.2
pytest-json==0.4.0
pytest-metadata==1.8.0
pytest-selenium==1.17.0
pytest-testconfig==0.2.0
pytest-variables==1.9.0
PyYAML==5.3
query-string==2019.4.13
redis==3.4.1
request==2019.4.13
requests==2.22.0
requests-toolbelt==0.9.1
selenium==3.141.0
setuptools-scm==3.5.0
six==1.14.0
SQLAlchemy==1.3.13
urllib3==1.25.8
wcwidth==0.1.7
